requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
django-parler==1.9.2
django-polymorphic==2.0.2
django-registration-redux==2.2
django-tag-parser==3.1
django-template-analyzer==1.6.1
future==0.16.0
html5lib==1.0.1
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
